extract load
