+++ src/Loader.tsx
@@ -1,22 +1,10 @@
-import AsyncStorage from '@react-native-async-storage/async-storage';
-import { useState, useEffect } from 'react';
 import { levelerMachine } from './machine';
+import { useLoadSnapshot } from './persist';
 import { Root } from './Root';
 import { View } from 'react-native';
 
 export function Loader() {
-  const [snapshot, setSnapshot] = useState(null);
-
-  useEffect(() => {
-    async function load() {
-      const data = await AsyncStorage.getItem(
-        levelerMachine.id,
-      );
-      const persisted = JSON.parse(data || "undefined");
-      setSnapshot(persisted);
-    }
-    load();
-  }, []);
+  const snapshot = useLoadSnapshot(levelerMachine.id);
 
   return (
     snapshot === null ? <View /> : <Root snapshot={snapshot} />
